이벤트 상수 및 캐시 TTL 위치 변경
이벤트 상수와 캐시 TTL을 utils/etc.ts로 이동하여 코드 구조를 개선했습니다.

주요 변경사항:
1. backend/server/src/const/cache-const.ts 파일 삭제
2. 이벤트 상수와 캐시 TTL을 utils/etc.ts로 이동
3. 관련된 서비스 파일들에서 이벤트 상수와 캐시 TTL의 import 경로 수정
