--- backend/server/src/const/cache-const.ts
@@ -1,9 +1,0 @@
-export const EVENTS = {
-    JOURNAL_CREATED: 'journal.created',
-    JOURNAL_DELETED: 'journal.deleted',
-    DOG_CREATED: 'dog.created',
-    DOG_UPDATED: 'dog.updated',
-    DOG_DELETED: 'dog.deleted',
-} as const;
-
-export const CACHE_TTL = 1000 * 60 * 60;
